add integration between gateway and inference services

+++ gateway-service/internal/model/inference.go
@@ -3,5 +3,5 @@ package model
 type InferenceResult struct {
 	DocumentID string
 	Label      string
-	Confidence float64
+	Confidence float32
 }
